Added instance data for forward lights.
Populates the shader variables for forward lights in a scene.

Populate function for dsSceneInstanceData now takes the
dsShaderVariableGroupDesc instance as a parameter. In the case of forward
light data, this is used to get the number of lights.

Changed dsSceneLightSet_findBrightestLights() to return the number of
lights rather than using an inout parameter. This is more natural to use,
and the loss of errno isn't very important since it should only fail for
invalid parameters.