Each model within a model node now has an array of draw ranges.
Rather than having multiple model entries for the different ranges, have a
single model entry with an array of draw ranges. This will make it easier
to create different variations of a model with different draw list
configurations, since it will be a strictly one to one relationship rather
than one to many.

Updated the tools version to the latest.