Added model node cloning with recnfiguration.
A list is provided with the name of the model within the node to
reconfigure and the new configuration, including the shader, material, item
list, and draw range. The extra item lists are also provided.

An example usage is to have a base model converted from the source, then
reconfiguring it for various situations with different sets of render
passes, such as forward vs. deferred rendering.