Fixes for documentation and node cloning.
Renamed name JSON element to baseName for model cloning. This avoids a
naming conflict for the name of the node itself and the node it's cloning.

Fixed various documentation errors and added documentation for scene
lighting types.

Removed duplicate scene lighting flatbuffer scripts, which was added from
a previous build with an incorrect namespace.