Prefer discrete GPUs.
Set the magic variables on Windows to prefer discrete GPUs on hybrid GPU
systems. This must be done in main.cpp for RenderBootstrap unit tests
and SDLMain.c, since it must be exported in the .exe itself. (i.e. it can't
be exported from renderer library if a shared library)

DS_MAIN_EXPORT is now set correctly for Android. The Config.h file wasn't
included, causing the define to not be properly set per platform.