Changes to do lighting in view space.
Doing lighting in view space is better for deferred rendering since it
allows you to drop the Z coordinate for the normal, assuming it points into
the screen. This frees up a color channel for the geometry buffers for
other parameters.

Changed worldInvTrans to worldViewInvTrans for instance transform data.
LightData for TestScene and forward light data now store the direction in
view space.