Use separate values for surface and default anti-alias samples.
Surface samples are intended for render surfaces and default samples for
offscreens and renderbuffers. This isn't exclusive, since offscreens and
renderbuffers that are used with render surfaces may need to use the
surface samples to work properly. This is intended to be used in cases
where anti-aliasing would be performed to offscreens before performing
post-processing effects, which don't require anti-aliasing on the final
render surface.

Use DS_SURFACE_ANTIALIAS_SAMPLES for the surface samples and
DS_DEFAULT_ANTIALIAS_SAMPLES for the default samples. Scene and view config
files can now set samples to "Surface" or "Default" for the corresponding
values, and it is now required rather than optional.

TestLighting now sets the default samples for the forward lighting buffers,
while the render surface stays at 1 sample.