Compare swapchain rather than render surface data in framebuffer.
This prevents an issue where re-creating the render surface will yield the
same pointer, causing it to fail to re-create the framebuffer for the
previously destroyed swapchain. The swapchain is guaranteed to be safe
since both will exist at the same time, which prevents the handles from
ever being the same values.

This replaces the previous workaround of re-creating the full window.