Optimized lighting.
Added function to get the diffuse light value for all light types. This
avoids branching, which murders performance on some GPUs. This is primarily
noticed on Intel, which has a significant performance improvement. (though
Intel integrated is still laughably slow)

Always use half vector for specular lights. This is based on the lighting
model rather than light type, so shouldn't be changed based on the light
type.

Adjusted scene for TestLighting to reduce overdraw. In practice this will
have relatively few state changes from material transitions, so it will
always be more optimal.