Fixes and cleanup.
processVertex() functions for vector draw shaders now return the position
instead of using an output parameter. Calls to GL_ADJUST_CLIP() are still
on a separate line to avoid multiple calls when the parameter is referenced
multiple times in the macro.

Use blend_op in place of alpha_blend_op for render states. This worked
because add is the default, but both color and alpha were intended to share
the same op.

Use ambient macros instead of directional macros in deferred light resolve
for the ambient geometry. These match the directional values, but were
added later when formalizing the ambient light.