Added functions to get view position from a clip position.
Added inverted projection matrix to the view transform uniforms. Added
dsGetViewPositionFromScreen() to get the screen position and depth and
dsGetViewPositionFromClip() to get from a vec3 clip position.

Added DS_CLIP_HALF_DEPTH macro to shaders to determine if the depth range
is [0, 1] or [-1, 1]. Re-arranged the internal uniform in OpenGL to allow
the clip range to be set, since it's determined at initialization time.

Added documentation for the built-in defines.