Fixes and optimizations for deferred lights.
Fixed spotlights to use the correct trig value to compute the geometry.
Prevents spotlight from being cut off in TestLighting.

Only perform lighting on pixels that have depth values >= the geometry
depth. (or <= for inverted depth) When combined with early fragment tests,
only pixels that can be lit will pay the cost of performing the lighting.

Unless DS_USE_DEPTH_CLAMP is defined, the z value is manually clamped. This
could potentially cause some artifacts at the far plane since it
effectively changes the slope of the geometry, but is expected to have
minimal impact in practice compared to the savings.