Fixes for deferred lighting on Android.
Set the screen size based on the rotated size. This allows the clip space
to be correctly computed based on gl_FragCoord. Also apply the rotation
when computing the clip coordinates.

Input attachment read now sets the VK_ACCESS_SHADER_READ_BIT for subpass
dependency access flags. This fixes reading depth/stencil attachments on
both Snapdragon and Mali, though according to the Vulkan docs shouldn't
be needed.

Fixed detection of when to keep attachments. Attachments with one sample
are only kept if the dsAttachmentUsage_KeepAfter bit is set. This should
reduce memory bandwidth significantly, especially on mobile platforms.

Use VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL for depth/stencil
attachments that are also used as an input attachment, This allows for
read-only usage.

Still broken on Mali, where only VK_IMAGE_LAYOUT_GENERAL seems to work.
Based on the Vulkan documentation this appears to be a driver bug. The only
way to work around this would be to use VK_IMAGE_LAYOUT_GENERAL for all
depth/stencil surfaces. For now assume that it's due to using old drivers,
given that it's still stuck on a 3 year old version of Android.