Added notes for broken Mali behavior.
After further investigation, using VK_IMAGE_LAYOUT_GENERAL doesn't work
around the issue. Most likely some other change made it seem like it was
working.

The key seems to be a combination of early fragment tests and performing
the subpass load. There's no point in removing the early fragment tests
since the depth is just for optimization, so it's recommended to simply
not bind the depth buffer on Mali if it's still present on target devices.