Made light shaders optional for dsDeferredLightResolve.
Can now set any combination of ambient, directional, point, and spot light
shaders to NULL. This allows for only processing a subset of light types,
such as in different render passes. For example, ambient can be split into
a separate render pass for SSAO, which requires samples outside of the
current pixel location.