Added option for unbalanced Kd tree.
Should be faster to build than a balanced tree at the expense for lookup
time.

The axis is now inferred from the depth. This allows for a smaller node
structure and avoids having to compute the bounds at each level for
balanced trees.

Added the option to stop the traversal from the Kd tree traversal function.