Added nearest neighbor function to Kd tree.
Searches for the single closest object in the Kd tree to the input point.

Root node is now explicitly stored rather than always the middle node. This
simplifies the multiple traversal functions as well as building an
unbalanced tree.

Removed unnecessary check in Kd tree balanced building for count being 0.
The checks for recursion explicitly cause this case to not happen.