Can now specify multiple shader/material/list triplets for models.
In scene resource JSON files, rather than having a single shader, material,
and model list for each piece of a model node, store multiple triplets in
a "modelLists" array. This avoids needing to duplicate each model entry,
which can be especially useful when multiple LODs are included. The
flatbuffer schema is the same, so it will create an entry in the model for
each modelList member.

Fixed manually created model pieces (rather than converting from a file
like OBJ or glTF) not being forwarded to the flatbuffer file. This still
hasn't been fully tested, but it should at least be more likely to work
now.