Minor fixes and cleanup.
Fixed incorrect errno and casts for implicit conversions that static code
analysis complained about. Added explanation for infinite planes for
dsFrustum3*.

Updated errno section of core README.md. Also added an explanation as to
why errno is used, not just to explain to others but also to remind myself
whenever I think that errno feels bad to use. (yes it is bad, but the added
API and usage complexity is arguably worse given that it's only for logging
info)