Added dsShadowProjection to compute shadow projection matrix.
Initialize with camera and light information, then add points to determine
the extents. Once all the points are added, the final projection matrix can
be computed.

The projection matrix is a light space projection matrix by default,
falling back to uniform (orthographic) shadows when the view vector is
parallel to the light direction or explicitly requested.

A light projection matrix may optionally be provided to handle spot and
point lights. When provided, the operations are computed in post-projected
space.