Kd trees are now always balanced.
Like normal binary trees, Kd trees will have increased depth if they are
unbalanced. In practice wih typical use cases, this increased depth makes
building more expensive compared to the sorts required to balance at each
level. Unlike a BVH, where the depth of the tree is fixed and being
unbalanced just means less optimal intermediate bounding boxes, literally
every part of the Kd tree from building to traversal and lookups is slower
when unbalanced, so it's not worth providing as an option for the contrived
case where the input is in exactly the right order to be balanced.