Added relative epsilon equal check.
This is more accurate than computing an epsilon ahead of time for
situations like shadow cull volumes. This avoids having a very far plane
causing points closer to the origin from having too loose of a check. For
example, having a far plane of 10000 won't end up using an epsilon of 0.1
for all points.

Added a test for non-intersecting frustums with a spot shadow cull volume.
There are currently remaining issues for intersecting frustums.