Implemented shadow cull volume intersect functions.
Performs a similar check to dsFrustum3f, then optionally adds the points of
the box or sphere to the shadow projection.

Since the volume should be guaranteed to be convex, the plane mask for
corners and edges are used to help clamp the points to the cull volume for
a tighter fitting projection. If no corner or edge matches, the point is
clamped to each plane it falls behind.