View now uses dsProjectionParams.
Now uses parameterized projection rather than a direct projection matrix.
This allows computations on the projection, namely for shadows to adjust
the near and far planes. When using a perspective projection, the aspect
ratio is automatically set.

Added separate functions for degree/radian conversion for float and
radians. This likely doesn't have any performance difference, but it avoids
the need for casting in many situations.