Adjusted scene light shadows.
All lights now respect the max shadow distance. Shadows will be disabled if
a point or light shadow is outside of this adjusted projection matrix, or
if the intensity is otherwise too low.

Removed the texture from dsSceneLightShdadows. This will need to be set on
the dsView instance to maintain the framebuffer.