Fixes for offscreens and framebuffers.
Only allow 3D texture offscreens on Vulkan if the maintenance1 extension
is provided. This is required to have a 2D Array compatible image view to
bind any layers beyond the first one.

Guarantee the first layer will always be used as the starting point when
multiple layers are bound. This is a requirement for OpenGL, which is
either all or one layer, while other implementations could break if it was
set to a non-zero value.

Better test out of bounds for framebuffer layers, as resolved offscreens or
3D textures weren't properly checking the internal layers.

Create a 2D array image view for 3D offscreens on Vulkan. This requires the
VK_IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT flag, which was added with the
maintenance1 extension. The resolve image and image view on Vulkan is now
always a 1D or 2D image without cube faces or arrays.