Added culling logic for scene light shadows.
Can cull for a specific surface to get the intersection result and add to
the shadow projection. Once all culling is done for a surface, the surface
projection can be computed. It's assumed culling between surfaces may be
multi-threaded, so atomics are used track which surfaces have been
computed.

Added accessors to mutate dsSceneLightShadows once it's created.