Added main light to dsSceneLightSet.
The main light will always be in index 0 of the brightest lights. Can also
apply shadows to the main light with an overload to
dsComputeForwardLighting(). Can query in the shader if the main light is
set with dsHasMainLight(), e.g. to avoid performing a shadow lookup if the
main light is out of view or unset.