Hooked up directional light shadows for deferred lighting in TestLighting.
Added shadow texture attribute for deferred light info. This looks up the
shadow texture based on the light shadow's name to hook up to the shader.

Fixed the index buffer count for deferred lights. This fixes having
multiple lights of the same type and shadowed lights, which are placed at
the end of the buffer.

Fixed reading of shadow light info for DeferredLightResolve when converting
JSON scene to flatbuffers.