Adjusted shadows.
Use a smaller epsilon for avoiding degeneracies when adding points for
projected shadow types. Avoids occasional artifacts for point shadows in
TestLighting.

Adjusted slope depth bias values for point and spot lights. These are tuned
for Vulkan, OpenGL seems to use a different computation for the slope and
would need to be adjusted separately. A constant scale doesn't appear to be
sufficient.