Avoid conditionals in shaders for clip differences.
Switched DS_CLIP_HALF_DEPTH macro to DS_MIN_CLIP_Z. This can be used
directly when clamping the depth value for reversed Z for deferred
lighting. In dsGetViewPositionFromClip(), math is used in OpenGL based on
DS_MIN_CLIP_Z to avoid a conditional, and is disabled for other targets.