Updated Android build.
Updated flatbuffers to the latest version.

Handle cases where 32-bit float depth isn't supported when loading scenes
from file. This is the case with at least some Mali GPUs.

Fixed lighting artifact on Snapdragon in the center of the screen due to
small half vectors. Need to clamp the length above 1e-3 to avoid this
artifact.