Fixed handling of cull volumes along shadow volume limits.
Rather than attempting to clamp the points to the volume, the segments for
the box are clamped to the volume planes. Additionally, the corners of the
volume inside the box are added. Fixes artifacts along edges, which are
more pronounced for tall screens. (e.g. phone in portrait orientation)

Added oriented box functions to check for point containment. Also added
ray evaluation function.