Adjusted how bounds are clamped to the shadow volume.
Only larger bounds (10% of the far plane) are clamped for directional
shadows. Still clamp for spot and point shadows to avoid precision issues,
especially since they tend to be more limited in scope and use smaller
textures.

Fixes occasional artifacts for very tall windows in TestLighting, such as
when running on a phone.