Avoid passing arrays to shader functions.
Passing array uniform members to a function that takes an array parameter
generates code that explicitly copies each individual array element to
another variable. This code isn't optimized out even when optimizations are
enabled, and makes the SSAO pass take a full second on Adreno.

Moved uniforms for SSAO, cascaded shadows, and point shadows to the headers
that declare them. For the shadows, defines can be used to control the
uniform and function names to allow them to be included multiple times if
multiple shadow lookups are required. Parameters are still used for the
generic dsShadowMap() function as it doesn't require arrays.