Switched back to full screen resolve for SSAO in TestLighting.
In practice the fragment shader does better than the compute shader for
texture lookup locality. Swizzling the ID to process the screen in columns
of a fixed width gives a marginal improvement, but still not as good as a
straight fragment shader. In cases where the values can be cached in shared
memory across the thread group, such as a blur with a fixed (or maximum)
kernel size, compute shaders can be faster, but SSAO doesn't fit those
parameters.