Added support for dedicated allocations in Vulkan.
Load the VK_KHR_dedicated_allocation extension when available to perform
dedicated memory allocations when the driver prefers it. This is essential
for render target memory on NVidia devices to avoid memory bandwidth
bottlenecks.

This fixes general slowdowns on NVidia cards, especially seen when
multisampling, and avoids random stutters. TestLighting, which is the only
tester that does non-trivial rendering, each lighting mode is now as fast
or faster in Vulkan compared to OpenGL.

Updated Vulkan headers to the latest.