Fixed point light shadow artifacts seen on iPad.
Instead of computing a min box size for shadows, explicitly set a min depth
range for each cascade. This sets the minimum distance between the near and
far planes for the shadow projection by setting the minimum size of the Y
axis of the box used to compute it.

This should make depth biases more consistent in general when few items are
in range for shadows. In this case it was likely device-specific
differences that had either less precision or different depth bias
behavior.