Added more color transform functions.
Added matrices to convert between YUV and RGB colors. This allows for
easier transforms for hue and saturation.

Added function to create an HSV color transform matrix. This allows
rotation of hue and scale of saturation and value with a linear transform.

These utilities are available both in C (through Color.h) and shaders
(through Color.mslh).