Hooked up game controller mappings to dsGameInput.
Added mapping value to game input button and axis events. This will be set
when available, otherwise the raw index will be set. Added functions
to get the axis and button states based on the mapping, converting as
needed if SDL treats it as an axis or button.

Hooked up controller touch events to populate the appropriate values and
pass the controller the events came from.