Added missing Application functionality.
Added function to get the current event time. Since event times are
relative to an implementation-specific start point, getting the current
time is often needed to make proper use of them.

Added function to get the current system power state. This includes a
high-level power state, including if there's a battery and if it's
charging, seconds remaining on battery, and remaining percentage. These
shouldn't be relied on for absolute accuracy, and is largely intended to
alert the user of low battery e.g. when playing in full screen.

Protect game input battery query with version check. Theoretically any
version of SDL2 should be supported.