Exposed scene tree hierarchy.
Can now traverse the hierarchy, get the underlying node, a pointer to the
transform, an the instance data. Allows for more elaborate usage of the
scene, including getting a parent transform or data from another node.

Add node function for dsSceneItemList now takes the tree node rather than
the transform. Removed dsSceneInstanceInfo in favor of passing the
dsSceneTreeNode pointers to the instance variable populate function. This
gives overall more information availabe in a more compact form.