Hooked up transform handling to dsSceneStandardParticleEmitterFactory.
Added update function to particle node and emitter factory. If set it will
be called in place of the standard emitter update function.

Standard emitter factory will set the transform to the relative node if
set, otherwise to the tree node's transform. The volume transform is
modified to be relative to the particle node.