Updated tools to version 24.
Now uses universal binaries for Mac tools, allowing the full build process
to be native on newer Macs with Apple CPUs.

Updated ModularShaderLanguage submodule. Now requires CMake 3.5.
Technically this shouldn't be necessary since 3.5 is required for the boost
targets, which won't be used for only the client library, but is old enough
that it isn't worth working around.

Updated README.md. Now documented that the default "mac" library package is
a universal binary and the architecture-specific Mac packages.