Fixes for cached shaders.
dsCreateDirectory() now returns the proper value.

Use pipelineCacheUUID from physical device properties for Vulkan to name
the pipeline cache file. This ensures that the proper cache is used when
the device changes, such as swapping out the GPU or toggling between
 available GPUs on the system. (e.g. dual Intel/NVIDIA laptops)

Save the shader cache for OpenGL when the shader is destroyed. This ensures
that the cache was warmed after usage, otherwise it may only have the bare
minimum of program data. It's now always written rather than only the first
time after a shader change to allow it to change over time. (e.g. new
driver version, new pipeline instantiations)

Fixed memory leak when saving cached shaders for OpenGL.

Scene resources are now destroyed in reverse order to creation. This is
more robust and should avoid any situations where resources are destroyed
in the wrong order compared to the two-pass destruction done before.