Fixed pipeline management in Vulkan.
Key based on the dsVkRenderPassData's lifetime object rather than
dsVkRenderPass. This ensures that if the underlying Vulkan pipeline needs
to be re-created the pipeline will be re-created as well. While
multisampling is part of the pipeline key, if the shader is used with a
subpass that doesn't use multisampling but the render pass as a whole does
it previously wouldn't see that the pipeline would need to be re-created.
For example, this was visible in TestLighting when toggling forward
lighting anti-aliasing.