Fixed unstable math when getting polygon edge angles.
Explicitly check for collinear edges with an epsilon and treat them the
same regardless of the ccw parameter. This avoids tiny floating point
values from massively changing the value and ambiguity for CW and CCW for
the degenerate case.

Fixes unit tests when when optimizations are enabled and targeting newer
architectures, which may have additional or different roundoff error
depending on the instructions used.