Use xoshiro256** for random number generation.
This is a new random number generator (from 2018) and is one of the best
available generators. It provides output that is statistically
indistinguishable from true random numbers for all known stress tests,
has a long period (2^256 - 1), has a reasonable state size of 4 64-bit
integers (32 bytes), and is very fast.

Added dsRandom structure to hold the state rather than just an integer
state and functions now follow the object convention of being prefixed by
dsRandom_. Added more explicit functions for getting the next random number
for signed and unsigned integers with an implicit minimum value of 0 vs. a
full range, normalized floating point values in the range of [0, 1), and
floating point values in an explicit range.

Unit tests for random number generation now do rudimentary checks to verify
that the results are well-distributed. No longer tests for exact values as
C++ random doesn't provide a builtin implementation to compare to.