Improved random number generation.
dsRandom_createSeed() now uses device random bits, falling back to the
previous behavior on failure.

Added dsRandom_nextSeed() to advance the seed directly. This avoids having
to create a dsRandom instance in situations such as
dsSceneStandardParticleEmitterFactory when a single seed is used to
initialize multiple dsRandom instances. Removes the need to have a full
dsRandom instance to advance the seed.

Added dsRandom_initialize() to fully initialize the dsRandom instance with
device random bits. If dsDeviceRandomBytes() failes or fills with the
invalid state of all 0s (next to impossible), it will fall back to a
time-based random seed.

Inline dsRandom_next(). This optimizes to 12 instructions on ARM64 and 20
instructions on x86_64 and would be dominated by function call overhead.
Given how simple the function is, there's no real cost to inline it.

Removed bias when requesting random integer ranges. Using a float to
compute spreads out the bias but doesn't remove it. This also allows the
full range to be used rather than 24 or 53 bits. Integer ranges are now
inclusive to allow the full range to be used.

Added dsRandom_nextSignedFloat() and dsRandom_nextSignedDouble() functions
to get a value in the range (-1, 1). Also added
dsRandom_nextFloatCenteredRange() and dsRandom_nextDoubleCenteredRange() to
get values in the range (center - range, cener + range). These allow for
symmetric ranges that doesn't bias towards the lower value.

Updated particle code to use the best random functions for their uses.