Added resource action type for scene resource files.
A resource action performs an action without having an actual resource
associated to it, though other resources might be created as a result of
the action.

Added TransformNodeChildren resource action to add children to a transform
node that's already been created. This is needed for TestParticles in order
to reference the root node for the particle factories as the relative
transform node. The children that reference those particle factories must
be added after the particle factories were declared.

Added particle emitter factories to scene graph for TestParticles.