Avoid double static linking with SDL.
Pre-built libraries now contain both dynamic and static SDL. Static SDL is
chosen when building DeepSea as static, otherwise dynamic SDL is used. The
appropriate library is forwarded for the install.

Copy over any DLLs to the output directory on Windows to allow them to be
found when running local builds.