Fixes for Android.
Explicitly set the main library name as metadata in the manifest. When
multiple targets are built, all .so libraries are copied rather than just
the one for the target being built. As a result, the wrong tester might
be invoked depending on the order of libraries in the package.

Don't load hidapi library as it's no longer built separately with SDL.

Disabled RenderFunctionalTest TextureBuffer test on ARM Mali GPUs for
OpenGL ES. This crashes when dispatching the compute shader on the Pixel 7.