Added dsSceneLightNode.
Holds a template light to create lights based on instances within the scene
graph. Creating the light is done with dsSceneLightSetPrepare.

dsSceneLightSetPrepare now only operates on a single light set. This makes
it possible to know witch light set to add lights for light nodes to
without having to store the light set in the node itself.