Cache full joint transforms for animation trees.
Converts to node local space before multiplying by the transform for the
joint. The inverse transpose is also computed for transforming directions.

Enforce that children of joint nodes are a higher index. This allows the
transforms to be updated in a single linear loop rather than a recursive
function.

Switched to storing animation node children with a pointer to the array
rather than an unsized struct array.