Use matrix box representation for cull nodes.
Allows the matrix to be used end to end. Shadow culling can get the corners
from the matrix, and added optimized SIMD path for expanding the shadow
bounds when not clamping.

Cull nodes can now store a static bounds for an optimized code path that
doesn't require calling the bounds function. This is used for model nodes
by default. Static and dynamic bounds are processed separately in cull
lists to allow for optimal code path for the most common case of static
bounds.