Added dsSceneAnimationNode.
Holds an animation tree and a shared list of direct and keyframe animation
maps. Once the appropriate item list type is added, the animation tree will
be cloned and a dsAnimation created to manage the animations.

Removed dsSceneAnimation as the current plan is to have the dsAnimation
instances be managed by the item list data stored in the dsSceneTreeNode
instances.