Added pre-transform update function to scene item lists.
Allows transforms to be updated within scene lists. This will be used for
updating animations and applying animations to scene nodes that are
connected to animation nodes.

dsScene_update() is now profiled. As more computation is done in the update
function this will help track the performance impact, especially as it is
single-threaded.