Added dsSceneAnimationTransformNode.
Contains the name ID of the animation node to apply the transform for.
dsSceneAnimationList will connect the transform and mark the node as dirty
when the value changes.

Made transform management more generic for dsSceneTreeNode. A baseTransform
pointer is now used for determining the transform. Added setupTreeNodeFunc
to optionally set up a tree node, which dsSceneTransformNode uses to set to
the node's transform.