Allow more flexibility for management of animations.
Added dsAnimationNodeMapCache to cache registration of direct and keyframe
animations with animation trees. This is now held with dsAnimation and
internally manages the node maps, which are no longer publicly exposed.
This allows the same animation to be used with multiple animation trees,
such as for different LODs. (e.g. high LOD with more bones than a low LOD)

Removed custom public facing scene animation and animation tree types as
these are now managed by dsAnimationNodeMapCache. Added
dsSceneAnimationTreeNode to allow multiple animation trees to be under a
dsSceneAnimationNode. Animation trees are now lazily applied in scenes to
allow for use cases like different LODs to save CPU cycles. The lazy
application is thread-safe.