Implemented glTF/GLB conversion for animation joint trees.
Similar to animation trees with joints, except it reads the skin elements
from glTF. The joint list is a flat list of indices and has a list of
matrices to transform to local space. Indices will be remapped as necessary
from the input glTF data and output flattened node structure.