Changed macros for starting SIMD blocks.
Use a comma-separated list for features. This allows multiple features to
be enabled more easily and works around an issue in Apple clang that only
pays attention to the first attribute push.

Due to preprocessor limitations and pragma requirements, spaces may not be
used between parameters in DS_SIMD_START().