Better handle custom alignments.
Added dsAllocator_alignedAlloc() to ensure the high-level allocator
interface supports aligned allocations.

dsBufferAllocator and dsPoolAllocator are now more lenient for alignment.
dsPoolAllocator will verify that the buffer and chunk size allow the
alignment, while dsBufferAllocator will offset the buffer as needed to
align the pointer. DS_REALIGNED_SIZE() can be used to determine the size
required to handle the worst-case realignment based on the default
allocation alignment.

Alignment is now enforced to be a power of two and optimized under that
assumption.