Fixed TestAnimation on Vulkan and OpenGL.
Fixed binding of vertex buffers on Vulkan. It was largely coincidence that
the previous situations happened to work.

Correctly bind integer vertex elements and shader uniform buffers on
OpenGL.

Added DS_ADJUST_TEXTURE_CLIP() for clip values that are guaranteed to be
part of a texture. This is used in place of DS_ADJUST_CLIP() for shadow
values, allowing for correct shadow lookups when drawing to the main window
framebuffer.

Adjusted TestAnimation lights and properly convert final color values to
sRGB.