Work around inconsistent floating point results on Mac ARM64.
When building on release, it gets different results even when performing
the exact same operation (dsVector2_len2(a)) twice for the two parameters
of EXPECT_EQ(). This behavior even happens assigning dsVector2_len(a) to
two separate variables and passing them to EXPECT_EQ(). Most likely the
compiler is optimizing them differently, resulting in a different order of
computations. Use floating point values that should avoid precision loss
through any permutation of how it orders the multiplies and adds.