Improvements for dsThreadPool.
Wait for the threads to finish starting or stopping in
dsThreadPool_setThreadCount(). Should guarantee that concurrent calls are
safe, or even calls on the same thread if the workers take too long to
start up.

Added dsThreadPool_getThreadCountUnlocked(). May be used to avoid locking
overhead or potential stalls when the workers are searching for tasks if
guaranteed that the thread count won't be changed concurrently. (generally
true when called from the main application thread)