Exposed ability to set double vs. triple buffering.
Changed vsync to an enum, which can be set to disabled, double buffer, or
triple buffer. When using Vulkan with FIFO presentation mode, this will
determin whether to use 2 or 3 images to tradeoff between performance and
latency.

Renamed doubleBuffer renderer bool with singleBuffer to better
differentiate with the vsync settings. In nearly no situation should it be
true, and is only supported by OpenGL. No longer call swap buffers when
single buffered.