Added integration between dsResourceManager and dsThreadPool.
Added optional functions that can be called on thread start and stop. Added
dsResourceManager_createThreadPool() that uses these to acquire and release
a resource context.

Renamed functions to create/destroy a resource context to acquire/release.
When not using a dummy implementation, it generally keeps a pool of
resource contexts that are re-used.