Improved cubic evaluation for animations.
Exposed matrices to convert between Bezier and Hermite parameters and cubic
polynomials. When a dsKeyframeAnimation is created, the input Hermite
tangents and positions are used to create the four cubic polynomials using
the Hermite to cubic matrix in a form that can evaluate all four components
with a 4x4 matrix transformation of the vector [1, t, t^2, t^3].

Removed redundant operations for rotations when applying animations.
Normalizing before applying the weight shouldn't be a large enough
difference justify the extra work, and no need to scale by the reciprocal
of the total weight before normalizing.