Force at least one recursion for non-linear cubic curves.
Some curves will pass through the midpoint and have the same tangent
direction as a line, which would avoid any recursions. Check the cubic
polynomial to verify if the t^2 or t^3 factors are non-zero, and if so
force at least one recursion. This situation should only occur at the
midpoint for cubics, other locations where the position check passes should
be caught by the tangent checks.