Allow item lists with global data to be processed in parallel.
View global values must now be locked with dsView_lockGlobalValues() and
unlocked with dsView_unlockGlobalValues(). Proper usage is enforced by
making the globalValues struct member const to avoid accidental
modification, while the functions to lock and unlock global values take
the dsSceneItemList instance to ensure it set globalValueCount to a value
> 1.

No longer require that item lists with global data are alone within its
sub-array for sharedItems in the scene. This allows slower item lists to
be run in parallel when setting global values, such as
dsSceneShdowManagerPrepare when there are many shadow surfaces. Only the
small portion that needs to set the values on globalValues will be locked.